Implement CooldownCircle in ItemSlotManager
Reorganize files
